+++ project_root/model/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, Integer, String, Boolean, BigInteger
 from sqlalchemy.orm import relationship
 
 from project_root.model import Base
@@ -7,7 +7,7 @@ from project_root.model import Base
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(String, unique=True, nullable=False)
     is_admin = Column(Boolean, nullable=False, default=False)
